Getting tracker to run
Switch to just grabbing closest state in lattice for tracking planner
and using the policy of that node.

Special policy for achieving the goal rotation: once the displacement
tolerance is satisfied, just keep rotating in the correct direction
to achieve it. Previous version would oscillate because it could
switch back and forth between policies of nodes that push the wheelchair
towards each other. This also eliminates the need to search over
motion primitives or do rollouts, the next config can be directly
found at each time step.

From previous commit: Updated grid planner so that it directly
penalizes going close to obstacles.

Works for all the test worlds.